requirements.txt
# -- Google ADK & Gemini -------------------------------------------------------
google-adk>=0.3.0
google-genai>=1.0.0

# -- Character AI unofficial API -----------------------------------------------
characterai>=1.0.0        # https://pypi.org/project/characterai/

# -- Web automation ------------------------------------------------------------
playwright>=1.44.0        # run: playwright install chromium

# -- Mobile automation ---------------------------------------------------------
Appium-Python-Client>=3.1.0
selenium>=4.20.0          # Appium client depends on selenium

# -- Utilities -----------------------------------------------------------------
python-dotenv>=1.0.0      # loads .env file
